Cleanup iteration for scripts and readme
The readme is out of date, so this change is
a start to bring it up to date and ensure it's
accurate.

The cleanup.sh script now wraps the optional "nukem all"
docker commands with a prompt to check. The total docker
nuke probably needs to go somewhere else completely, maybe
into a dot file.

+++ cleanup.sh
@@ -17,7 +17,13 @@ done
 
 # This is a sort of nuclear option, and will more or less delete everything
 # which isn't running.
-docker system prune -af && \
-    docker image prune -af && \
-    docker system prune -af --volumes && \ # deletes build cache objects
-    docker system df # shows disk usage
+infotext "Do you want to nuke all docker containers and images?"
+read nukem
+if [[ $nukem == "yes" ]]; then
+  docker system prune -af && \
+      docker image prune -af && \
+      docker system prune -af --volumes && \ # deletes build cache objects
+      docker system df # shows disk usage
+else
+  infotext "Not nuking docker containers and images."
+fi
